Fix tests and remove swagger doc
